requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
paddleocr==2.7.3
# Padrão CPU (troque para GPU conforme Dockerfile.gpu)
paddlepaddle==2.6.1
pytest==8.3.2
slowapi==0.1.9
python-multipart==0.0.9